python/submission_report_job/requirements.txt
Jinja2
python-json-logger
python-keycloak
psycopg2-binary
pytz
requests
responses